+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import './index.css';
-import Game from './Game';
+import Game from './components/Game';
 
 
 ReactDOM.render(
